[refactor] extract RowItem component
