+++ src/screens/SopDetailScreen.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
-import { Icon } from 'native-base'
+import { Container, Button, Content, Header, Left, Right, Body, Title, Icon } from 'native-base'
 import moment from 'moment';
 import { typeIcon } from '../config/utils'
 
@@ -9,11 +9,26 @@ const SopDetailScreen = props => {
   let { type, icon, color } = typeIcon(sopGuide.document_type)
 
   return (
-    <View style={styles.screen}>
-      <Icon type={type} style={{ color, fontSize:30 }} name={ icon } />
-      <Text>{ sopGuide.tags }</Text>
-      <Text>created date: { moment(sopGuide.created_at).fromNow() }</Text>
-    </View>
+    <Container>
+      <Header>
+        <Left>
+          <Button onPress={ () => props.navigation.goBack() }>
+            <Icon name="md-arrow-back" />
+          </Button>
+        </Left>
+        <Body>
+          <Title>Sop detail</Title>
+        </Body>
+        <Right />
+      </Header>
+      <Content padder>
+        <View style={styles.screen}>
+          <Icon type={type} style={{ color, fontSize:30 }} name={ icon } />
+          <Text>{ sopGuide.tags }</Text>
+          <Text>created date: { moment(sopGuide.created_at).fromNow() }</Text>
+        </View>
+      </Content>
+    </Container>
   );
 };
 
@@ -37,4 +52,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default SopDetailScreen;
+export default SopDetailScreen
